feature/adminSalesCateAnalasysA#123 완성
추가 및 개선사항
> seller의 코드를 이용 최대한 빠르게 만들었음

+++ BeetMall/src/main/java/com/beetmall/sshj/seller/vo/SellerSalesVO.java
@@ -18,7 +18,7 @@ public class SellerSalesVO {
 	
 	// 차트에 들어갈 데이터?
 	private int[] resultData; // 뷰단에 존재하는 가져와야 할 데이터 목록[배열]
-	private int ordernum; // 주문번호
+	private long ordernum; // 주문번호
 	private String productname; // 상품이름(제목)
 	private int productnum; // 상품번호 
 	private int orderquantity; // 수량
@@ -201,10 +201,12 @@ public class SellerSalesVO {
 	public void setResultData(int[] resultData) {
 		this.resultData = resultData;
 	}
-	public int getOrdernum() {
+
+	
+	public long getOrdernum() {
 		return ordernum;
 	}
-	public void setOrdernum(int ordernum) {
+	public void setOrdernum(long ordernum) {
 		this.ordernum = ordernum;
 	}
 	public int getProductnum() {
